requirements.txt
django
djangorestframework
djangorestframework-simplejwt
django-filter
gunicorn==20.1.0
psycopg2-binary==2.9.1